requirements.txt
streamlit
Pillow
agno
duckduckgo-search